feat: Redesign Infrastructure, Virtual Machines Digital Twin Workspace, Live Telemetry, & Universal Data Ingestion Platform

+++ backend/src/main/java/com/bpp/digitaltwin/telemetry/LocalTelemetryCollector.java
@@ -121,7 +121,7 @@ public class LocalTelemetryCollector implements TelemetryCollector {
 
         // Data Quality Metadata
         event.quality = new DataQualityDto(true, 0, 100.0, "GOOD");
-        event.quality.latencyMs = m.networkLatencyMs;
+        event.quality.latencyMs = (long) m.networkLatencyMs;
 
         return event;
     }
